bundle images as a separate step
