+++ lib/block-list.js
@@ -34,25 +34,7 @@ function renderItem (item) {
 }
 
 export default function ({body}) {
-  const bundlesToUrls = {};
-  const urlsToBundles = {};
-  let count = 0;
-
-  const components = body
+  return body
     .map(renderItem)
     .filter(Boolean);
-
-  components.forEach(item => {
-    if (item.components && item.components[0] && item.components[0].role === 'photo') {
-      const url = item.components[0].URL;
-      if (!urlsToBundles[url]) {
-        urlsToBundles[url] = `image-${count}`;
-        bundlesToUrls[`image-${count}`] = url;
-        count++;
-      }
-      item.components[0].URL = `bundle://${urlsToBundles[url]}`;
-    }
-  });
-
-  return {components, bundlesToUrls};
 }
