+++ lib/index.js
@@ -2,6 +2,7 @@ import assert from 'assert';
 import renderBlockList from './block-list';
 import objectAssign from 'object-assign';
 import defaultStyles from './default-styles';
+import bundleImages from './bundle-images';
 
 function convertToAppleNews (articleJson, opts) {
   assert(opts, 'opts required');
@@ -10,7 +11,8 @@ function convertToAppleNews (articleJson, opts) {
   assert(articleJson.title, 'articleJson.title required');
   assert(Array.isArray(articleJson.body), 'articleJson.body required to be an array');
 
-  const {components, bundlesToUrls} = renderBlockList(articleJson, opts);
+  const components = renderBlockList(articleJson, opts);
+  const bundlesToUrls = bundleImages(components);
   const article = {
     version: '1.0',
     identifier: opts.identifier,
